[FEAT/#134] 트랜잭션과 롤백 정책 처리
